1329: re-factor react table
