+++ packages/client/src/Views/PlayerLog/styles.module.scss
@@ -96,24 +96,6 @@
   margin-right: 5px;
 }
 
-.filter-group {
-  width: 100%;
-  display: flex;
-  justify-content: flex-end;
-  margin-bottom: 5px;
-}
-
-.filter-icon {
-  position: absolute;
-  margin-top: 8px;
-  color: "#3ef1ea";
-}
-
-.input-filter {
-  color: #000;
-  width: 250px;
-}
-
 .custom-cell {
   display: flex;
   align-items: center;
