+++ packages/components/src/index.ts
@@ -45,4 +45,5 @@ export { default as CollabStatusBoard } from './local/collab-status-board'
 export { default as HeartbeatChecker } from './local/atoms/heartbeat-checker'
 export { default as Confirm } from './local/atoms/confirm'
 export { default as Tabs } from './local/atoms/tabs'
-export { Row, Column } from './local/collab-status-board/types/props'
+export { default as ReactTable } from './local/react-table'
+export { Row, Column } from './local/react-table/types/props'
