+++ packages/components/src/local/collab-status-board/styles.module.scss
@@ -4,44 +4,3 @@
   background-color: $background-action-head;
   padding: 10px;
 }
-
-.actions {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-}
-
-.filter-icon {
-  position: absolute;
-  margin-top: 8px;
-  color: "#3ef1ea";
-}
-
-.input-filter {
-  color: #fff !important;
-  width: 250px;
-  padding-left: 22px;
-}
-
-.checkbox {
-  margin-left: 0 !important;
-  margin-bottom: 0.2rem !important;
-
-  span {
-    color: #3ef1ea;
-  }
-}
-
-.btn {
-  span {
-    width: 100%;
-    display: flex;
-    justify-content: flex-end;
-  }
-  button {
-    color: #3ef1ea;
-    &:hover {
-      color: #fff;
-    }
-  }
-}
