+++ packages/components/src/local/collab-status-board/types/props.d.ts
@@ -1,22 +1,12 @@
 import {
-  MessageCustom,
-  ForceRole,
   ChannelCollab,
   ForceData,
+  ForceRole,
+  MessageCustom,
   TemplateBodysByKey
 } from '@serge/custom-types'
-import React from 'react'
-import { TableColumnBase } from 'react-data-table-component/dist/src/DataTable/types'
 
-export type Collapsible = (cb?: (() => void) | undefined) => React.ReactElement
-export type Row = { [x: string], collapsible?: Collapsible }
-export type Column = TableColumnBase & {
-  selector: (row: Row2) => any
-  sortFunction?: (rowA: Row, rowB: Row) => number
-  colFilter?: boolean
-}
-
-export default interface Props {
+export type CollabStatusBoardProps = {
   currentWargame: string
   messages: MessageCustom[]
   channelColb: ChannelCollab
