+++ packages/components/src/local/react-table/README.md
@@ -1,0 +1,27 @@
+# React Table documentation
+
+This is an example readme file for the 'React Table' component.
+
+## UI Example
+
+<!-- STORY -->
+
+## Code example
+
+<!-- SOURCE -->
+
+## Props
+
+<!-- PROPS -->
+
+## Usage
+
+### ES6 Import
+```js
+import ReactTable from './react-table'
+```
+
+### CommonJS
+
+```js
+const ReactTable = require('./react-table')
