+++ packages/components/src/local/react-table/styles.module.scss.d.ts
@@ -1,7 +1,13 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'rfi-form': string;
+  'actions': string;
+  'checkbox': string;
+  'btn': string;
+  'input-filter': string;
+  'filter-icon': string;
+  'light': string;
+  'dark': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
