+++ packages/components/src/local/react-table/types/props.d.ts
@@ -1,0 +1,18 @@
+import { TableColumn, TableProps } from 'react-data-table-component'
+
+export type Column = TableColumn & {
+  // true if we want to enable filter value for this column
+  colFilter?: boolean
+}
+export type Row = { [x: string], collapsible?: Collapsible }
+
+export type ReactTableProps = Omit<TableProps, 'columns'> & {
+  columns: Column[]
+  rows: Row[]
+  customStyles: any
+  showArchived?: boolean
+  handleArchiveDoc?: () => void
+  handleMarkAllAsRead?: () => void
+  channelName?: string
+  filterTheme?: 'dark' | 'light'
+}
